UI tests implimented and pass
