+++ app/src/androidTest/java/com/example/photoapp/ExampleInstrumentedTest.java
@@ -1,26 +1,36 @@
 package com.example.photoapp;
+import androidx.test.rule.ActivityTestRule;
 
-import android.content.Context;
-
-import androidx.test.platform.app.InstrumentationRegistry;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
-
+import org.junit.Rule;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
-import static org.junit.Assert.*;
+import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.action.ViewActions.click;
+import static androidx.test.espresso.action.ViewActions.closeSoftKeyboard;
+import static androidx.test.espresso.action.ViewActions.replaceText;
+import static androidx.test.espresso.action.ViewActions.typeText;
+import static androidx.test.espresso.assertion.ViewAssertions.matches;
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
+import static androidx.test.espresso.matcher.ViewMatchers.withText;
 
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-@RunWith(AndroidJUnit4.class)
 public class ExampleInstrumentedTest {
+    @Rule
+    public ActivityTestRule<MainActivity> mActivityRule = new ActivityTestRule<>(MainActivity.class);
     @Test
-    public void useAppContext() {
-        // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
-        assertEquals("com.example.photoapp", appContext.getPackageName());
+    public void ensureTextChangesWork() {
+        //Automated UI Test for search Activity
+
+        onView(withId(R.id.filter)).perform(click());
+        onView(withId(R.id.etStartDateTime)).perform(replaceText("20210120 000000"), closeSoftKeyboard());
+        onView(withId(R.id.etEndDateTime)).perform(replaceText("20210126 000000"), closeSoftKeyboard());
+        onView(withId(R.id.etKeywords)).perform(typeText("Potato"), closeSoftKeyboard());
+        onView(withId(R.id.go)).perform(click());
+        onView(withId(R.id.captionid)).check(matches(withText("Potato")));
+        onView(withId(R.id.nextid)).perform(click());
+        onView(withId(R.id.nextid)).perform(click());
+        onView(withId(R.id.previd)).perform(click());
+        onView(withId(R.id.previd)).perform(click());
+
     }
 }
+
